requirements.txt
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
coloredlogs==15.0.1
flatbuffers==23.5.26
humanfriendly==10.0
idna==3.6
imageio==2.33.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
lazy_loader==0.3
llvmlite==0.41.1
mpmath==1.3.0
networkx==3.2.1
numba==0.58.1
numpy==1.26.3
onnxruntime==1.16.3
opencv-python-headless==4.9.0.80
packaging==23.2
pillow==10.2.0
platformdirs==4.1.0
pooch==1.8.0
protobuf==4.25.2
PyMatting==1.1.12
referencing==0.32.1
rembg==2.0.53
requests==2.31.0
rpds-py==0.17.1
scikit-image==0.22.0
scipy==1.12.0
sympy==1.12
tifffile==2023.12.9
tqdm==4.66.1
urllib3==2.1.0
